requirements.txt
boto3==1.9.101
dj-database-url==0.5.0
Django==2.0.5
django-colorfield==0.1.15
django-cors-headers==2.4.0
django-heroku==0.3.1
django-jet==1.0.8
django-modeltranslation==0.13b1
django-solo==1.1.3
django-storages==1.7.1
django-tinymce==2.7.0
djangorestframework==3.9.0
gunicorn==19.9.0
pillow==5.4.1
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pytz==2018.9
whitenoise==4.1.2
